backend/requirements.txt
asgiref==3.3.4
astroid==2.5.6
Django==3.2.4
django-cors-headers==3.7.0
djangorestframework==3.12.4
isort==5.9.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
